requirements.txt
Django~=3.1.5
djangorestframework~=3.12.2
django-cors-headers~=3.6.0
pandas~=1.2.0
jupyter~=1.0.0
ipython~=7.19.0
django-extensions~=3.1.0